Complete the process in which are prepared the inactive processes and successively to be executed
In ProcessManagerWindow class:

Add a border width in the window.

Set the "timeout_id" property by a null value at start the program and use the timeout_add when is actived a switch, and at the same time show a spinner to indicate that is executing the simulation. Otherwise when the switch is disabled the "timeout_id" is stopped.

When is prepared a process in the "process_manager" and the limit of "prepared_processes" count is reached, then shows a error message by a MessageDialog that indicates the error.

Get and show the process when is deactivated or suspended by the methods "deactivate_proess_action" and "suspend_process_action" respectively.

In ProcessManager class:

Set the default value of property "current_pid" by the integer zero.

Calculate the maximum lenght of all the names in the "inactive_processes" property to set the padding for the names in the processes.

Change the way in which is prepared a process, create a new process with the name received by parameter in the "prepare_process" method

Return the process when is deactivated or suspended in the methods "deactivate_process" and "suspend_process" respectively.

In Process class:

Add the "name_pad" property, zero by default, is used for the paddig on some properties when it calls the __str__ method.

Change the name of the "priority_level" property to "priority", is changed to shorten the property name.

Add the __eq__ method to search processes when are removed from processes lists.