Add a new paramether to show all the process data
When is passed a True value at __str__ method, it returns a string with
 all process data, otherwise when is passed a False value the method
only returns the name and the pid of the process.

the new method paramether is called 'all' and has a False value by
default. The additional data is the priority, memory, quantum and
interaction of a process.