Change the spin button for quantum rat to miliseconds
Set the adjustement in thte quantum rat spin button to work with values
in the range from 250 miliseconds (0.25 s) to 10000 (10 s) and
increment each 250 miliseconds.

When is changed the quantum rat in the spin button is also updated in
the process manager, so it need divide the miliseconds by 1000 to
convert them to seconds.

Fix the iteration algoritm to pass the suspended process to prepared
state. First verifying if the process at 0 position in suspended
processes list has the 'next' attribute, as true remove the 'next'
attribute and move the process to prepared processes list. Otherwise
is verifyied the 'interaction' attribute of process, as true add the
'next' attribute to indicate that this process at next iteration
will be moved to prepared processes list. Otherwise is removed from
the suspended processes and added to prepared processes.

Also it was changed the way to suspend or deactivate a executed
process, always increase the progress when the execute is finished,
but it verify if the progress is completed, in this case the
process has finished completly and is deactivated.