Add a new attribute 'advance'
It was interchanged the value for 'processor_time' and 'quantum'
attributes, now the 'quantum' depends to 'processor_time'.

The new attribute 'advance' has the value (1 / quantum) and indicates
the percentage of advance for the process and it will be increase the
'progress attribute' when is moved from executed state to suspended
state.