Change the PRIORITIES dict to a tuple without values
Make the priority random choise without specify the dict keys.

Change the list to interaction random choise to a tuple.

Change the parameter 'all' in __str__ method to 'show_all'. Change the
string to return, remove the memory and add the processor time, round
the quantum and truncate the progress importing the 'trunc' function
from utilities module.