Implement the expropiation
Set at simulating_switch active attribute 'True' as default.

Move the initialization for the inactive processes buttons from
update_components to init_components method and remove the Toggle
functionality and active state.

Add remove_list_box method to remove from list_boxes the removed
processes in lists.

In execute_simulation method change the parameter name from gparam to
'_', and remove the timeout_id attribute always, only set a value when
the switch is active.

In add_process_action method change the parameter name from button to
'_' and add a button to inactive_processes_list_box directly.

Remove unused conditional in prepare_process_action method.

Change the unused parameter 'button' to '_' at iteration method and
specify the return type. Change the return value to boolean, returns
True when the expropiated_process attribute has not a object and False
when the value is None.

Call the execute_simulation method when is changed the
expropiated_process attribute, to change the method to be iterated
and have different types of execution.

Delete the value of expropiated_process attribute when is assigned to
None, also print a message to view the process executed at
expropiation.